Merge2

+++ code/modules/power/tesla/energy_ball.dm
@@ -131,17 +131,17 @@ var/list/blacklisted_tesla_types = typecacheof(list(/obj/machinery/atmospherics,
 	dust_mobs(A)
 
 /obj/singularity/energy_ball/orbit(obj/singularity/energy_ball/target)
-	if (istype(target))
+	if(istype(target))
 		target.orbiting_balls += src
 		poi_list -= src
 	. = ..()
 
 /obj/singularity/energy_ball/stop_orbit()
-	if (orbiting && istype(orbiting.orbiting, /obj/singularity/energy_ball))
+	if(orbiting && istype(orbiting.orbiting, /obj/singularity/energy_ball))
 		var/obj/singularity/energy_ball/orbitingball = orbiting.orbiting
 		orbitingball.orbiting_balls -= src
 	..()
-	if (!loc)
+	if(!loc && !QDELETED(src))
 		qdel(src)
 
 /obj/singularity/energy_ball/proc/dust_mobs(atom/A)
@@ -189,13 +189,14 @@ var/list/blacklisted_tesla_types = typecacheof(list(/obj/machinery/atmospherics,
 		else if(closest_grounding_rod || is_type_in_typecache(A, blacklisted_tesla_types))
 			continue
 
-		else if(istype(A, /mob/living))
-			var/dist = get_dist(source, A)
+		else if(isliving(A))
 			var/mob/living/L = A
-			if((dist < closest_dist || !closest_mob) && L.stat != DEAD)
-				closest_mob = L
-				closest_atom = A
-				closest_dist = dist
+			if(!L.tesla_ignore)
+				var/dist = get_dist(source, A)
+				if((dist < closest_dist || !closest_mob) && L.stat != DEAD)
+					closest_mob = L
+					closest_atom = A
+					closest_dist = dist
 
 		else if(closest_mob)
 			continue
